v0.1.4
기본 공격 로직 완성

+++ Assets/Scripts/animalID.cs
@@ -6,5 +6,5 @@ public class animalID : MonoBehaviour
 {
     public int Attack;
     public int Heart;
-    public int UnitID;
+    public float UnitID;
 }
